add registry services for texture and buffer managerment, setup renderer to manager rendering

+++ Engine/src/core/features/Mesh.h
@@ -21,6 +21,6 @@ struct Vertex {
 
 struct Mesh {
 	std::vector<Vertex> vertices;
-	std::vector<uint32_t> indices;
-	std::vector<uint32_t> materialIDs;
+	std::vector<uint16_t> indices;
+	std::vector<uint8_t> materialIDs;
 };
